gm_card_app_backend/requirements.txt
click==7.1.1
dnspython==1.16.0
Flask==1.1.2
itsdangerous==1.1.0
more-itertools==8.2.0
packaging==20.3
py==1.8.1
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.1
python-dotenv==0.13.0
Werkzeug==1.0.1
